Create notion_redirect_cloudflare.js
This works for viewing the homepage site. Also will allow you to click through links.
I had a problem with vivaldi browser which was resolved by updating line 92 to include current chromium versions as well as mozilla 5.0 for windows.
Now, the only problem is that the inline databases don't show up on pages that have them. So, this v1 is the preservation of the work I've gotten done so far.